Syncing with parent reo. (#1)
* Fix selector of the _order field to make stacked inlines sortable.

* Catch TinyMCE errors when trying to save inlines.

Previously this used to prevent the new _order values from being computed.

* Annotate special handling of the _order field in inlines.

* Add CSS to hide the _order field in stacked inlines.

* Stop using deprecated template settings.

- TEMPLATE_DIRS
- TEMPLATE_LOADERS

I think these must have been actually removed in django-1.10.0 because
they're throwing AttributeError's now.

* Bump dependencies.

* Remove deprecated richtext_filter tag.

I don't see how the fallback could even work without the user defining
RICHTEXT_FILTERS as empty in their own settings, because it defaults to
`("mezzanine.utils.html.thumbnails",)` so the `if not filternames` path
would never be taken. Give that the fallback does nothing, I think
printing a warning is deceptive so better just to remove it entirely.

* Exclude .pyc, .pyo and __pycache__ files from release

* Fixed crash in Page.get_ascendants with non-current site

* Add override_current_site_id context manager

* Make sure a thread-local request is set in PagesTests

We rely in these tests on a “current request” being stored as a
thread-local, but without explicitly assigning one we were dependent on
other tests that make requests being executed before these.

* Remove redundant assignment of request.site_id

* Correctly handle custom email fields in mezzanine.forms

The `is_a` method checks whether a field is one of Mezzanine's built-in
form fields. As far as I can tell, every other usage is for the purpose
of initializing the built-in fields and widgets, so it makes sense to
exclude user-defined fields.

However in this instance, we only want to know whether the field is an
EmailField, not that it is *the* built-in Mezzanine email field.
Therefore, check the class rather than identity.

The effective change here is that user-defined fields (in
FORMS_EXTRA_FIELDS) which subclass django's EmailField will be able to
receive confirmation emails.

* Add compatibility with the Django 1.10 MIDDLEWARE setting.